transfer modal content to separate components
