+++ src/components/burger-constructor/BurgerConstructor.tsx
@@ -2,9 +2,10 @@ import React, { useState } from 'react';
 import './BurgerConstructor.css';
 import { ConstructorElement }  from '@ya.praktikum/react-developer-burger-ui-components';
 import { Button } from '@ya.praktikum/react-developer-burger-ui-components';
-import { CurrencyIcon, CheckMarkIcon } from '@ya.praktikum/react-developer-burger-ui-components'
+import { CurrencyIcon } from '@ya.praktikum/react-developer-burger-ui-components'
 import { Counter }  from '@ya.praktikum/react-developer-burger-ui-components';
 import Modal from "../modal/Modal";
+import OrderInfo from './OrderInfo';
 
 interface Ingredient {
     _id: string;
@@ -52,13 +53,7 @@ const BurgerConstructor: React.FC<BurgerIngredientsProps> = ({ ingredients }) =>
             </div>
             {isModalOpen && (
               <Modal onClose={closeModal}>
-                <h1 className="text text_type_digits-large">034536</h1>
-                <h2 className="pt-8 pb-5 text text_type_main-medium">идентификатор заказа</h2>
-                <div className="pt-5 pb-5 modal-icon-wrap">
-                    <CheckMarkIcon type="primary" />
-                </div>
-                <p className="pt-5 pb-2 text text_type_main-default">Ваш заказ начали готовить</p>
-                <p className="text text_type_main-default text_color_inactive">Дождитесь готовности на орбитальной станции</p>
+                <OrderInfo/>
               </Modal>
             )}
         </section>
